Use @Inject, replace product component annotations
The use of the UserDirectoryService caused problems in the
confluence-confapi-plugin. This was probably caused by the missing @Inject
annotations and by the product component annotations (like @JiraComponent) as
these are only normal @Component annotations that are also annotated with a
e.g. @OnlyInProduct(ProductFilter.JIRA) annotation. The behaviour for multiple
products is not really clear.

And as these warnings now popped up: The deprecated MockitoJUnitRunner has now
been updated to the current version in the org.mockito.junit package.